requirements.txt
finnhub-python>=2.4.18
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
lxml>=4.9.0        # required by pandas.read_html for Wikipedia fetch
pytest>=7.4.0
pytest-mock>=3.12.0
